Remove Layer type
Previously, the `Map` type had a list of `Layer` types. For the current
map, I had a land layer and a sea layer. Since I always want only those
two layers, I decided to make them explicit keys on the map instead. I
also need access to the land layer for pathfinding.

At that point, the `Layer` type was just extra abstraction for nothing
so I removed it and inlined the lists of tiles in the `Map`.