Introduce global and local coordinate spaces
All entities are positioned absolutely on the map in **global** space
where (0,0) is in the bottom-left corner of the map.

However, movement and coordinate math for entities is easer to do in
**local* space where (0,0) is at the entity's current location.

This introduces `Coordinate.Global` and `Coordinate.Local` types to add
some safety and to distinguish between the two types of coordinates. It
also adds some helpers to make conversions easier.