Use a 0-based tile numbering system
The math for converting back and forth from tile number to (x, y)
coordinates becomes simpler if the first tile of the map (north-west
corner) is tile 0 rather than tile 1.

This also creates a new type to get extra safety around the use of the
tile number. It also helps guarantee I didn't miss any places that
needed to convert to 0-based.